Fix runtime error caused by passing DateTime object to select_date method
Whenver you call 'select_date' method with DateTime as a param, it throws and runtime error, because it tries to compare Date with DateTime at the end of a method call.
This results in failing conditional check and throwing an error.

This fix changes value we're operating on at the very beginning to make sure we always operate on Date object.